Stub out methods needed for Matrix class
Honestly, I could've stubbed out almost three times as many methods, but I need to limit the scope for my intended purposes, and not indulge my desire for 'completeness'. I can return and expand the classes in post-1.0.0 releases and slowly inch it towards what I desire.

Next up is to start implementing the actual methods in the Vector class.