Implement randomized constructor for Vector along with unit test
Did extensive testing and it works as desired at present. There could be some concerns with certain targets as I don't believe all of the will allow 'jumping' a parameter by providing null when it has a default value but isn't nullable according to the function signature, but its not something I care to look into any further at the moment (and I did look into it, but documentation is spartan and scattered - I'll have to write a test class in my scratchpad repo to tinker with it and various targets at some point).

Next up is to continue implementing Vector methods.