Implement internal and fromArray constructors for Matrix class, along with a unit test
Also learned one can make top-level declarations of any kind private to hide them from other modules in the same package, which I had to resort to avoid shenanigans with the FloatVector typedef. Knowing that Haxe actually allows one to break free from having things in classes (unlike Java), there is probably a better way for me to structure modules (like Assert) which are just meant to export functions to other modules in the same package, but I'll leave it as-is for now since it isn't a problem at the moment.

Next up is to crank out some more of the various Matrix constructor static methods.