Implement randomized Matrix constructor; extract out Haxe vector randomizer; write unit test
Passes. Refactored Vector to make use of the update function.

Again, I might run into issues with null passing and default argument values on certain targets, but I'm not going to concern myself with this now. The questions is where should the default values be specified, since min and max get chained through three functions. Will eventually write an HXML script that builds and runs for each target to identify where this will be an issue, but at a later time.

Next up is to continue writing the remaining methods for the Matrix class, along with corresponding unit tests.