Implement matrix-to-matrix math methods and row and column fetching methods
Nothing has been tested yet, and I am expecting bugs. Must test carefully, as this is a critical section of the code. I could've abstracted the Matrix as a collection of Vectors, but that would've added a lot of indirection and overhead if I scale this (which I intend to do in the programs that will make use of this library), hence why I keep using bare indices and raw access to the underlying Haxe vectors.

Next up is to write unit tests for the matrix to matrix math methods and the for the row and column fetching methods.