Implement unit tests for remaining methods
Finished implementing all desired tests = all are passing.

This library is now ready for an early release.

Next up is to prepare the project for upload to haxelib, polishing the README and various other miscellaneous tasks. I'm now ready to start writing the other libraries that will build off of this linear algebra library.