Expand testing to all targets; write script to build as importable Lua module
With a bit of guidance from other Haxe devs over on the forums, I was able to learn that I could just write names of classes on HXML lines without any preceding flag for inclusion, and that I could use each and next to chain together different build steps.

Next up is to rename the test invocation script, then introduce an invocation script that does it for just the eval / interpreted target, and then update documentation, and then publish an initial version.